// react-app/src/misc/openapi.d.ts
export interface components {
    schemas: {
// ...
        SourceItemResponse: {
            id?: string | null;
            link?: string | null;
            title?: string | null;
            description?: string | null;
            image?: string | null;
            showImage?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}